fix(player): strengthen preparingSongId guards, handle manual seek skip & re-publish master on cancelNext

+++ app/src/main/java/com/unshoo/pixelmusic/data/remote/youtube/YoutubeHelper.kt
@@ -1321,7 +1321,7 @@ object YoutubeHelper {
             suffix.contains("webm") || name.contains("webm") -> "audio/webm"
             else -> null
         }
-        Triple(selectedStream.content, mime, selectedStream.averageBitrate.toInt())
+        Triple(selectedStream.content, mime, selectedStream.averageBitrate)
     }
 
     // ════════════════════════════════════════════════════════════════════════════════════
